requirements.txt
pymysql
cryptography
psycopg2-binary
valkey
valkey[hiredis]